+++ src/main/java/de/mrstein/customheads/stuff/GetHistory.java
@@ -46,15 +46,15 @@ public class GetHistory {
 
     public Inventory getInventory() {
         Inventory hisInv = Bukkit.createInventory(null, 9 * 5, CustomHeads.getLanguageManager().HISTORY_INV_TITLE.replace("{PLAYER}", offlinePlayer.getName()));
-        hisInv.setItem(0, CustomHeads.getTagEditor().setTags(new ItemEditor(Material.STAINED_GLASS_PANE, (short) 15).setDisplayName(CustomHeads.getLanguageManager().HISTORY_SEARCHHISTORY).getItem(), "history", "open#>search"));
-        hisInv.setItem(1, new ItemEditor(Material.STAINED_GLASS_PANE, (short) 5).setDisplayName(CustomHeads.getLanguageManager().HISTORY_GETHISTORY_ACTIVE).getItem());
-        hisInv.setItem(8, new ItemEditor(Material.SKULL_ITEM, (short) 3).setDisplayName("§b" + offlinePlayer.getName()).setOwner(offlinePlayer.getName()).getItem());
+        hisInv.setItem(0, CustomHeads.getTagEditor().setTags(new ItemEditor(Material.STAINED_GLASS_PANE,  15).setDisplayName(CustomHeads.getLanguageManager().HISTORY_SEARCHHISTORY).getItem(), "history", "open#>search"));
+        hisInv.setItem(1, new ItemEditor(Material.STAINED_GLASS_PANE,  5).setDisplayName(CustomHeads.getLanguageManager().HISTORY_GETHISTORY_ACTIVE).getItem());
+        hisInv.setItem(8, new ItemEditor(Material.SKULL_ITEM,  3).setDisplayName("§b" + offlinePlayer.getName()).setOwner(offlinePlayer.getName()).getItem());
         for (int i = 9; i < 18; i++) {
-            hisInv.setItem(i, new ItemEditor(Material.STAINED_GLASS_PANE, (short) 1).setDisplayName("§0").getItem());
+            hisInv.setItem(i, new ItemEditor(Material.STAINED_GLASS_PANE,  1).setDisplayName("§0").getItem());
         }
         if (hasHistory()) {
             for (int i = 0; i < entries.size(); i++) {
-                hisInv.setItem(i + 18, new ItemEditor(Material.SKULL_ITEM, (short) 3).setDisplayName("§a" + entries.get(i)).setLore(CustomHeads.getLanguageManager().HISTORY_GET_LORE).setOwner(entries.get(i)).getItem());
+                hisInv.setItem(i + 18, new ItemEditor(Material.SKULL_ITEM,  3).setDisplayName("§a" + entries.get(i)).setLore(CustomHeads.getLanguageManager().HISTORY_GET_LORE).setOwner(entries.get(i)).getItem());
             }
         } else {
             hisInv.setItem(31, new ItemEditor(Material.BARRIER).setDisplayName("§0").setLore(CustomHeads.getLanguageManager().HISTORY_NO_HISTORY_LORE).getItem());
