+++ src/main/java/de/mrstein/customheads/stuff/SearchHistory.java
@@ -46,11 +46,11 @@ public class SearchHistory {
 
     public Inventory getInventory() {
         Inventory hisInv = Bukkit.createInventory(null, 9 * 5, CustomHeads.getLanguageManager().HISTORY_INV_TITLE.replace("{PLAYER}", offlinePlayer.getName()));
-        hisInv.setItem(0, new ItemEditor(Material.STAINED_GLASS_PANE, (short) 5).setDisplayName(CustomHeads.getLanguageManager().HISTORY_SEARCHHISTORY_ACTIVE).getItem());
-        hisInv.setItem(1, CustomHeads.getTagEditor().setTags(new ItemEditor(Material.STAINED_GLASS_PANE, (short) 15).setDisplayName(CustomHeads.getLanguageManager().HISTORY_GETHISTORY).getItem(), "history", "open#>get"));
-        hisInv.setItem(8, new ItemEditor(Material.SKULL_ITEM, (short) 3).setDisplayName("§b" + offlinePlayer.getName()).setOwner(offlinePlayer.getName()).getItem());
+        hisInv.setItem(0, new ItemEditor(Material.STAINED_GLASS_PANE,  5).setDisplayName(CustomHeads.getLanguageManager().HISTORY_SEARCHHISTORY_ACTIVE).getItem());
+        hisInv.setItem(1, CustomHeads.getTagEditor().setTags(new ItemEditor(Material.STAINED_GLASS_PANE,  15).setDisplayName(CustomHeads.getLanguageManager().HISTORY_GETHISTORY).getItem(), "history", "open#>get"));
+        hisInv.setItem(8, new ItemEditor(Material.SKULL_ITEM,  3).setDisplayName("§b" + offlinePlayer.getName()).setOwner(offlinePlayer.getName()).getItem());
         for (int i = 9; i < 18; i++) {
-            hisInv.setItem(i, new ItemEditor(Material.STAINED_GLASS_PANE, (short) 1).setDisplayName("§0").getItem());
+            hisInv.setItem(i, new ItemEditor(Material.STAINED_GLASS_PANE,  1).setDisplayName("§0").getItem());
         }
         if (hasHistory()) {
             for (int i = 0; i < entries.size(); i++) {
